+++ intro/fibo-array4.c
@@ -37,35 +37,35 @@
    I save the results in a list, and then I copy the list to an array
 */
 int main(int argc, char **argv) {
-        uint64_t limit = atoi(argv[1]);
-        uint64_t size = 0;
-        uint64_t* results = fibo_array(&size, limit);
-        for (uint64_t n = 0; results[n] > 0; n++)
-                printf("%" PRIu64 ": %" PRIu64 "\n", n + 1, results[n]);
+        uint32_t limit = atoi(argv[1]);
+        uint32_t size = 0;
+        uint32_t* results = fibo_array(&size, limit);
+        for (uint32_t n = 0; results[n] > 0; n++)
+                printf("%" PRIu32 ": %" PRIu32 "\n", n + 1, results[n]);
         free(results);
 }
 
 /**
    \brief actual function
 */
-uint64_t* fibo_array(uint64_t* sp, uint64_t limit) {
+uint32_t* fibo_array(uint32_t* sp, uint32_t limit) {
         GSList* list = NULL;
-        uint64_t prev = 1;
-        uint64_t prev2 = 1;
-        uint64_t actual = 0;
+        uint32_t prev = 1;
+        uint32_t prev2 = 1;
+        uint32_t actual = 0;
         if (limit >= 1)
                 list = g_slist_append(list, GINT_TO_POINTER(1));
         if (limit >= 2)
                 list = g_slist_append(list, GINT_TO_POINTER(1));
-        for (uint64_t n = 2; actual <= limit; n++) {
+        for (uint32_t n = 2; actual <= limit; n++) {
                 actual = prev + prev2;
                 list = g_slist_append(list, GINT_TO_POINTER(actual));
                 prev2 = prev;
                 prev = actual;
         }
         *sp = g_slist_length(list);
-        uint64_t* results = malloc(*sp * sizeof(uint64_t));
-        uint64_t i = 0;
+        uint32_t* results = malloc(*sp * sizeof(uint32_t));
+        uint32_t i = 0;
         for (GSList* p = list; p != NULL; p = g_slist_next(p), i++) {
                 results[i] = GPOINTER_TO_INT(p->data);
         }
